fix: Only display batch execute in queue, add tooltip to button
